Nicholas - Commit 6
Instead of having multiple methods to generate and build the file, it is
now done in one method called writeFile in IOFormatter.

Made the representation of data part of the DatabaseAccessor class to
streamline the process, thus NetworkBuilder is no longer used. This is
to prevent lessen the amount of 'getter' methods.