CR-16 - DCN and DAN generation: Modifying criteria
The following menu item
Choose:
1. Include Developers Who Comments On At Least One Bug
2. Only Include Bug Owners
Please Enter Your Choice (1 or 2):

Has been changed as:
Choose:
1. Include developers who comment on at least one bug in RESOLVED status
2. Include developers who own at least one bug in RESOLVED status and
comment on at lease one bug in RESOLVED status.
Please Enter Your Choice (1 or 2):

DCN has to be generated only for bugs in RESOLVED status. DAN has be to
be generated only for bugs in RESOLVED status. Generate DAN only for the
vertices of DCN.